forgot to move sfs.py to /sfs
